Update uber.cpp
I updated it so the process will run through every step if the user could not be passed, and had to choose either debit or bankaccount for funds. the uber call function simply gets its location, asks for the destination, and then couts that the driver is on its way, with 10 dollars being decucted.